+++ LibraryLocationQuerySystem/Pages/Locations/Delete.cshtml.cs
@@ -22,7 +22,7 @@ namespace LibraryLocationQuerySystem.Pages.Locations
         [BindProperty]
         public Location Location { get; set; } = default!;
 
-        public async Task<IActionResult> OnGetAsync(byte? LocationLevel, short? LocationId)
+        public async Task<IActionResult> OnGetAsync(byte? LocationLevel, int? LocationId)
         {
 			if (LocationLevel == null || LocationId == null || _context.Location == null)
 			{
@@ -43,7 +43,7 @@ namespace LibraryLocationQuerySystem.Pages.Locations
             return Page();
         }
 
-        public async Task<IActionResult> OnPostAsync(byte? LocationLevel, short? LocationId)
+        public async Task<IActionResult> OnPostAsync(byte? LocationLevel, int? LocationId)
         {
 			if (LocationLevel == null || LocationId == null || _context.Location == null)
 			{
