+++ LibraryLocationQuerySystem/Pages/Stores/Index.cshtml.cs
@@ -55,11 +55,11 @@ namespace LibraryLocationQuerySystem.Pages.Stores
 
         public class SelectGroupViewModel
         {
-            public short CampusId { get; set; }
-            public short LibraryId { get; set; }
-            public short FloorId { get; set; }
-            public short BookshelfId { get; set; }
-            public short LayerId { get; set; }
+            public int CampusId { get; set; }
+            public int LibraryId { get; set; }
+            public int FloorId { get; set; }
+            public int BookshelfId { get; set; }
+            public int LayerId { get; set; }
         }
 
         public IList<Store> StoreList { get; set; } = default!;
